pyproject.toml
[project]
name = "aiosellers"
version = "0.1.0"
description = "All In One Sellers API - Universal API for various digital trading platforms (such as PlayerOK or FunPay). Provides developers with a convenient way to create bots on any platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "idna>=3.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "wrapper-tls-requests>=1.1.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
# Keep linting minimal + style-oriented; formatting is handled by `ruff format`.
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
